requirements.txt
google-genai
python-dotenv
fastapi
uvicorn
python-multipart
sqlalchemy
jinja2
